Added some static_casts to avoid compiler warnings.
--HG--
extra : convert_revision : svn%3Afdd8eb12-d10e-0410-9acb-85c331704f74/trunk%403838

+++ dlib/matrix/lapack/geev.h
@@ -210,7 +210,7 @@ namespace dlib
                 return info;
 
             if (work.size() < work_size)
-                work.set_size(work_size, 1);
+                work.set_size(static_cast<long>(work_size), 1);
 
             // compute the actual decomposition 
             info = binding::geev(jobvl, jobvr, n, &a(0,0),
